Added authenticated and unauthenticated pages test files. Added more tests for model creation. Updated fakers to support having users that must have evaluations assigned to test evaluation length
